requirements.txt
requests
typing
tqdm
transformers
pillow
datasets